fix splash screen logic
