+++ packages/manager/src/components/AuthenticationWrapper/AuthenticationWrapper.tsx
@@ -23,6 +23,8 @@ import { handleLoadingDone } from 'src/store/initialLoad/initialLoad.actions';
 import { State as PendingUploadState } from 'src/store/pendingUpload';
 import { MapState } from 'src/store/types';
 
+import { SplashScreen } from '../SplashScreen';
+
 interface Props {
   children: React.ReactNode;
 }
@@ -72,8 +74,6 @@ export class AuthenticationWrapper extends React.Component<CombinedProps> {
       !this.state.showChildren
     ) {
       this.makeInitialRequests();
-
-      return this.setState({ showChildren: true });
     }
 
     /** basically handles for the case where our token is expired or we got a 401 error */
@@ -90,8 +90,12 @@ export class AuthenticationWrapper extends React.Component<CombinedProps> {
   render() {
     const { children } = this.props;
     const { showChildren } = this.state;
-    // eslint-disable-next-line
-    return <React.Fragment>{showChildren ? children : null}</React.Fragment>;
+
+    if (!showChildren) {
+      return <SplashScreen />;
+    }
+
+    return children;
   }
 
   static defaultProps = {
@@ -142,6 +146,7 @@ export class AuthenticationWrapper extends React.Component<CombinedProps> {
       /** We choose to do nothing, relying on the Redux error state. */
     } finally {
       this.props.markAppAsDoneLoading();
+      this.setState({ showChildren: true });
     }
   };
 }
