--- packages/manager/src/components/SplashScreen/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SplashScreen';
