+++ packages/manager/src/index.tsx
@@ -10,7 +10,6 @@ import { Route, BrowserRouter as Router, Switch } from 'react-router-dom';
 import AuthenticationWrapper from 'src/components/AuthenticationWrapper';
 import { CookieWarning } from 'src/components/CookieWarning';
 import { Snackbar } from 'src/components/Snackbar/Snackbar';
-import SplashScreen from 'src/components/SplashScreen';
 import 'src/exceptionReporting';
 import Logout from 'src/layouts/Logout';
 import { setupInterceptors } from 'src/request';
@@ -18,6 +17,7 @@ import { storeFactory } from 'src/store';
 
 import { App } from './App';
 import { LinodeThemeWrapper } from './LinodeThemeWrapper';
+import { SplashScreen } from './components/SplashScreen';
 import { loadDevTools, shouldEnableDevTools } from './dev-tools/load';
 import './index.css';
 import { queryClientFactory } from './queries/base';
@@ -36,58 +36,7 @@ const LoginAsCustomerCallback = React.lazy(
 );
 const OAuthCallbackPage = React.lazy(() => import('src/layouts/OAuth'));
 
-const NullAuth = () => <span>null auth route</span>;
-
 const Null = () => <span>null route</span>;
-
-const AppWrapper = () => (
-  <>
-    <SplashScreen />
-    <Snackbar
-      anchorOrigin={{ horizontal: 'right', vertical: 'bottom' }}
-      autoHideDuration={4000}
-      data-qa-toast
-      hideIconVariant={true}
-      maxSnack={3}
-    >
-      <App />
-    </Snackbar>
-  </>
-);
-
-const ContextWrapper = () => (
-  <ReduxStoreProvider store={store}>
-    <QueryClientProvider client={queryClient}>
-      <LinodeThemeWrapper>
-        <CssBaseline />
-        <React.Suspense fallback={<SplashScreen />}>
-          <Switch>
-            <Route component={OAuthCallbackPage} exact path="/oauth/callback" />
-            <Route
-              component={LoginAsCustomerCallback}
-              exact
-              path="/admin/callback"
-            />
-            {/* A place to go that prevents the app from loading while refreshing OAuth tokens */}
-            <Route component={NullAuth} exact path="/nullauth" />
-            <Route component={Logout} exact path="/logout" />
-            <Route component={CancelLanding} exact path="/cancel" />
-            <AuthenticationWrapper>
-              <Switch>
-                <Route component={Lish} path="/linodes/:linodeId/lish/:type" />
-                <Route component={AppWrapper} />
-              </Switch>
-            </AuthenticationWrapper>
-          </Switch>
-        </React.Suspense>
-      </LinodeThemeWrapper>
-      <ReactQueryDevtools
-        initialIsOpen={false}
-        toggleButtonProps={{ style: { marginLeft: '3em' } }}
-      />
-    </QueryClientProvider>
-  </ReduxStoreProvider>
-);
 
 const Main = () => {
   if (!navigator.cookieEnabled) {
@@ -95,13 +44,54 @@ const Main = () => {
   }
 
   return (
-    <Router>
-      <Switch>
-        {/* A place to go that prevents the app from loading while injecting OAuth tokens */}
-        <Route component={Null} exact path="/null" />
-        <Route component={ContextWrapper} />
-      </Switch>
-    </Router>
+    <ReduxStoreProvider store={store}>
+      <QueryClientProvider client={queryClient}>
+        <LinodeThemeWrapper>
+          <CssBaseline />
+          <Router>
+            <React.Suspense fallback={<SplashScreen />}>
+              <Switch>
+                <Route
+                  component={OAuthCallbackPage}
+                  exact
+                  path="/oauth/callback"
+                />
+                <Route
+                  component={LoginAsCustomerCallback}
+                  exact
+                  path="/admin/callback"
+                />
+                {/* A place to go that prevents the app from loading while refreshing OAuth tokens */}
+                <Route component={Null} exact path="/null" />
+                <Route component={Logout} exact path="/logout" />
+                <Route component={CancelLanding} exact path="/cancel" />
+                <AuthenticationWrapper>
+                  <Switch>
+                    <Route
+                      component={Lish}
+                      path="/linodes/:linodeId/lish/:type"
+                    />
+                    <Snackbar
+                      anchorOrigin={{ horizontal: 'right', vertical: 'bottom' }}
+                      autoHideDuration={4000}
+                      data-qa-toast
+                      hideIconVariant={true}
+                      maxSnack={3}
+                    >
+                      <Route component={App} />
+                    </Snackbar>
+                  </Switch>
+                </AuthenticationWrapper>
+              </Switch>
+            </React.Suspense>
+          </Router>
+        </LinodeThemeWrapper>
+        <ReactQueryDevtools
+          initialIsOpen={false}
+          toggleButtonProps={{ style: { marginLeft: '3em' } }}
+        />
+      </QueryClientProvider>
+    </ReduxStoreProvider>
   );
 };
